+++ packages/core/CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.6.0] - 2020-04-07
 ### Added
 - `AbstractVideoIngestionAppliance` now generates payloads with timestamps, and accepts an `origin` setting on appliance construction.
 
@@ -58,7 +60,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Increase the `base-classes` version dependency to `1.3.0`.
 - Increase the `base-interfaces` version dependency to `3.0.0`.
 
-[Unreleased]: https://github.com/tvkitchen/appliances/compare/@tvkitchen/appliance-video-file-ingestion@0.5.0...HEAD
+[Unreleased]: https://github.com/tvkitchen/appliances/compare/@tvkitchen/appliance-video-file-ingestion@0.6.0...HEAD
+[0.6.0]: https://github.com/tvkitchen/appliances/releases/tag/@tvkitchen/appliance-core@0.6.0
 [0.5.0]: https://github.com/tvkitchen/appliances/releases/tag/@tvkitchen/appliance-core@0.5.0
 [0.4.0]: https://github.com/tvkitchen/appliances/releases/tag/@tvkitchen/appliance-core@0.4.0
 [0.3.1]: https://github.com/tvkitchen/appliances/releases/tag/@tvkitchen/appliance-core@0.3.1